Finalizado toda la creación de las clases DAO, empezamos con la parte de la fachada

+++ tarea2DWES24/src/main/java/dao/OperacionesCRUD.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 
 public interface OperacionesCRUD<T> {
 
-	public int insertar( T elemento);
+	public long insertar(T elemento);
 
 	public T buscarPorID(long id);
 
@@ -13,5 +13,5 @@ public interface OperacionesCRUD<T> {
 	public boolean modificar(T elemento);
 
 	public boolean eliminar(T elemento);
-
+	
 }
